Put vector3 extensions in a separate file.
